Stylize App, optimize UX

+++ main.py
@@ -1,12 +1,13 @@
 # Kivy
 from kivy.app import App
-from kivy.uix.boxlayout import BoxLayout
 from kivy.properties import ObjectProperty
 from kivy.properties import StringProperty
 from kivy.properties import BooleanProperty
 from kivy.uix.popup import Popup
+from kivy.uix.button import Button
+from kivy.graphics import Color, RoundedRectangle
 from kivy.clock import Clock
-from kivy.uix.screenmanager import ScreenManager, Screen
+from kivy.uix.screenmanager import ScreenManager, Screen, SlideTransition
 # Other
 from threading import Thread
 import traceback
@@ -15,34 +16,28 @@ import os
 from app import GoogleTTS, MP3Player, loadPDF
 
 
-# Intro splash screen
+"""Intro splash screen"""
 class SplashScreen(Screen):
-    # once User enters the app
+    # splash screen appears for 3 seconds
     def on_enter(self, *args):
         Clock.schedule_once(self.switch_to_main, 3)
 
-    # switch to main screen
+    # Eventually switch to main screen
     def switch_to_main(self, *args):
-        # Switch to the main screen
         self.manager.current = "main"
 
 
-# The file selector popup
+"""The file selector popup"""
 class FileSelectorPopup(Popup):
     # File loader Object
     load = ObjectProperty(None)
 
-    def load(self, path, selection):
-        if selection:
-            print(f"Selected file(s): {selection}")
-            print(f"From path: {path}")
-        else:
-            print("No file selected.")
 
-# Main screen
+"""Main screen"""
 class MainScreen(Screen):
+    # Labels
     file_path = StringProperty("No file selected")
-    main_text = StringProperty("Please select a PDF file")
+    file_select_state = StringProperty("Tap the button below to choose a PDF file and generate an audiobook.")
 
     def open_file(self):
         # Open the file selector popup
@@ -50,65 +45,124 @@ class MainScreen(Screen):
         # Set the load callback with the popup instance
         popup.load = lambda path, selection: self.load_file(path, selection, popup)
         popup.open()
+    
+    # Temporarily disabled
+    # # Dynamic Button
+    # def create_audiobook_button(self):
+    #     # Get the container and buttons
+    #     audiobook_button_container = self.ids.audiobook_button
+    #     choose_file_button = self.ids.choose_file_button
 
+    #     # Avoid duplicates
+    #     if any(isinstance(child, Button) for child in audiobook_button_container.children):
+    #         return
+
+    #     # Dynamic button creation
+    #     audiobook_button_layout = Button(
+    #         text='Audiobook Player',
+    #         size_hint_x=1,
+    #         pos_hint={'center_x': 0.5, 'center_y': 0.5},
+    #         background_color=(0, 0, 0, 0), # blue background (#1C4983)
+    #         background_normal='',
+    #         color=(232/255, 241/255, 255/255, 1),  # Light text (#e8f1ff)
+    #         padding=[24, 16],
+    #         font_size="20sp",
+    #         font_name="./app/Fonts/Roboto_SemiCondensed-Medium.ttf"
+    #     )
+
+    #     # Create new canvas
+    #     audiobook_button_layout.canvas.before.clear()
+    #     with audiobook_button_layout.canvas.before:
+    #         Color(46/255, 93/255, 159/255, 1) 
+    #         rectangle = RoundedRectangle(pos=audiobook_button_layout.pos, 
+    #                          size=audiobook_button_layout.size, 
+    #                          radius=[15, 15, 15, 15])
+        
+    #     # bind position and size to the button and text
+    #     audiobook_button_layout.bind(pos=lambda instance, value: setattr(rectangle, 'pos', value))
+    #     audiobook_button_layout.bind(size=lambda instance, value: setattr(rectangle, 'size', value))
+        
+    #     # add button to container
+    #     audiobook_button_container.add_widget(audiobook_button_layout)
+
+    #     # adjust size hint
+    #     choose_file_button.size_hint_x = 0.5
+    #     audiobook_button_container.size_hint_x = 0.5
+
+    #     # onpress listener
+    #     audiobook_button_layout.on_press = self.switch_to_screen
+        
+
+    # Load the selected PDF file
     def load_file(self, path, selection, popup):
         print("Loading file")
+
+        # Check if selection is empty
         if not selection:
             print(f"No file selected from {path}")
             return
 
         try:
-            # Load the selected PDF file
+            # Create audiobook button once PDF selected
+            # self.create_audiobook_button()
+
+            # Change string property after selecting path
+            full_file_path = os.path.join(path, selection[0])
             self.file_path = selection[0]
+            self.file_select_state = "Load another PDF file to generate an audiobook."
             print(f"Attempting to load: {self.file_path}")
             
-            # dismiss popup first
+            # Dismiss selector popup
             popup.dismiss()
             
-            # extract text
+            # extract text from pdf
             pdf_content = loadPDF(self.file_path)
             if not pdf_content:
                 print("Error: Could not extract text from PDF")
                 return 
-                
-            # display text
-            self.pdf_text = pdf_content
+            # Change Label
+            fix_file_path = full_file_path.replace('\\', '/')
+            self.file_path = f"Recent file: {fix_file_path.split('/')[-1] if '/' in fix_file_path else fix_file_path.split('//')[-1]}"
+
+
+            # display text count
             print(f"Successfully loaded PDF with {len(pdf_content)} characters")
 
-            # changing text from TextScreen to PDF content
+            # changing text string property from TextScreen to PDF content
             text_screen = self.manager.get_screen("text")
             text_screen.text = pdf_content
             text_screen.has_pdf = True
-
-            # check if mp3 exists in google cloud bucket
 
             # remove old mp3
             if os.path.exists("output.mp3"):
                 os.remove("output.mp3")
 
             # Switch to text screen
-            self.manager.current = 'text'
+            self.switch_to_screen()
+
         except Exception as e:
             print(f"Error loading file: {str(e)}")
             traceback.print_exc()
 
     # switch to text screen
     def switch_to_screen(self, *args):
+        self.manager.transition = SlideTransition(direction='left')
         self.manager.current = "text"
 
 
-# Loading waiting popup
+"""Loading waiting popup"""
 class loading_tts_popup(Popup):
     pass
 
-# Pdf viewer
+
+"""PDF viewer"""
 class TextScreen(Screen):
-    text = StringProperty("")
-    # objects
+    # Instances
     google_tts = GoogleTTS()
     mp3_player = MP3Player()
-    # popup = loading_tts_popup()
+
     # booleans
+    text = StringProperty("")
     has_pdf = BooleanProperty(False)
     play_button_disabled = BooleanProperty(True)
 
@@ -117,11 +171,20 @@ class TextScreen(Screen):
         self.popup = loading_tts_popup()
 
     # checks if mp3 exists, Used for the kivy lang and create_tts function
-    def check_file_exists(self):
+    def check_audio_exists(self):
         if os.path.exists("output.mp3"):
             return True
         else:
             return False
+    
+    def check_thread_done(self, dt):
+        if not self.play_thread.is_alive():
+            self.popup.dismiss()
+            self.play_button_disabled = False
+            self.play_audio()
+            print("Audio created, audio playing")
+            return False  # stops the clock schedule
+        return True  # keeps checking
     
     # Creates mp3 from google tts API
     def create_tts(self):
@@ -130,7 +193,7 @@ class TextScreen(Screen):
         self.play_button_disabled = True
 
         # creates mp3
-        if not self.check_file_exists():
+        if not self.check_audio_exists():
             try:
                 # tell user that audio is being created
                 self.popup.open()
@@ -146,15 +209,6 @@ class TextScreen(Screen):
                 traceback.print_exc()
                 self.popup.dismiss()
                 self.play_button_disabled = False
-
-    def check_thread_done(self, dt):
-        if not self.play_thread.is_alive():
-            self.popup.dismiss()
-            self.play_button_disabled = False
-            self.play_audio()
-            print("Audio created, audio playing")
-            return False  # stops the clock schedule
-        return True  # keeps checking
 
     # plays mp3 using vlc
     def play_audio(self):
@@ -178,25 +232,37 @@ class TextScreen(Screen):
 
     def switch_to_main(self, *args):
         self.stop_audio()  # Stop audio before switching screens
+        self.manager.transition = SlideTransition(direction='right')
         self.manager.current = "main"
 
 
-# Builds App
+"""Initialize App"""
 class AudioBookApp(App):
     def build(self):
-        self.load_kv("styles.kv")  # Load the kv file explicitly
+        # Window properties
+        from kivy.config import Config
+        Config.set('graphics', 'resizable', '0')
+        Config.set('graphics', 'width', 812)
+        Config.set('graphics', 'height', 600)
+
+        # load KV File
+        self.load_kv("styles.kv")
+
+        # screens
         sm = ScreenManager()
         sm.add_widget(SplashScreen(name="splash"))
         sm.add_widget(MainScreen(name="main"))
         sm.add_widget(TextScreen(name="text"))
         return sm
     
+    # Once the app is either closed, or exited
     def on_stop(self):
         print("App stopped")
-        os.remove("output.mp3")
-        print("output.mp3 deleted")
+        if os.path.exists("output.mp3"):
+            os.remove("output.mp3")
+            print("output.mp3 deleted")
         print("Bye")
 
-
+"""Runs app"""
 if __name__ == '__main__':
     AudioBookApp().run()
